pyproject.toml
[project]
name = "feature-lens"
version = "0.1.0"
description = "In-depth visualizations for SAE features"
authors = [{ name = "Daniel Tan", email = "dtch1997@users.noreply.github.com" }]
dependencies = [
    "pre-commit>=3.7.1",
    "transformer-lens>=2.2.2",
    "sae-lens>=3.12.1",
    "eindex-callum>=0.1.2",
    "circuitsvis>=1.43.2",
    "torch>=2.1.2",
    "ipykernel>=6.29.5",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
test = ["ruff>=0.4.3", "pytest>=8.2.0", "pyright>=1.1.361"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.ruff]
line-length = 88
lint.select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I",
]
exclude = ["third_party", ".venv"]
lint.ignore = ["F841", "E501", "E741", "I001", "E731", "F722"]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["pdm-python-project-template"]
force-single-line = true
single-line-exclusions = ["typing"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["pdm-python-project-template"]

[tool.semantic_release]
version_variables = [
    "feature_lens/__init__.py:__version__",
    "pyproject.toml:version",
]
branch = "main"
